Trim leading whitespace from tmpl strings
JQuery 1.9+ has trouble with leading whitespace when creating
DOM from jQuery:

     // Bad in 1.9+ :
     $("    <div></div>")

     // Fine everywhere:
     $("<div></div>")

But the existing tmpl function was creating HTML strings with leading
whitespace. Fix in tmpl() implementation to trim it, jQuery is happy
again.

Fixes #17

This PR also includes a newly generated lib/jquery.stackview.min.js
That's messy, if multiple PR's exist at once there are going to be
conflicts, but I don't see any other way to do it if the generated
minimized js is in the repo.