Make all types respect linking to item.link
webpage, videofilm, and soundrecording each used their own
stacklife-backend-specific methods of assigning the href
in the item link.

For general use, it makes more sense for all types to respect
the common item.link.  I think I added it in a way that will
be backwards compatible, the other behavior is still there
if item.link is undefined and/or other special attributes are,
depending on the prior implementation. But I don't have an easy
way to test with your existing backend to ensure my assumptions
about backwards-compat were correct.