Fix styling of .stack-pages:after
The new styling matches what I reverse engineered from
the stacklife demo, and has slightly different border
color/width/style then was actually in the stackview repo.

What was in stacklife demo looks a lot better. So this PR.

Is stacklife using a fork of this stackview code? Possibly
with various improvements? Is there any easy way for me to diff
between what stacklife is actually using and stackview, and see
if there are any other missing improvements?